DMP-4910 Always set case closed date on case closed event
